Fix name for Rust completion
When using rust-analyzer for completion, the function is split up such that
the name is stored in word, e.g. `VimComplete.word = do_something`
however only the structure is stored in info/menu
e.g. `VimComplete.menu = fn(self, u64, ...)`. In this case we take the same
approach as go abstract functions and override the name.